requirements.txt
playwright==1.35.0
numpy == 1.25
opencv-python == 4.8.0.74
Pillow == 10.0.0
pytesseract == 0.3.10
python-dotenv == 1.0.0
PyPDF2 == 3.0.1